requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile ../requirements/requirements.in
#
asgiref==3.4.1
    # via django
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.6
    # via requests
django==3.2.7
    # via
    #   -r ../requirements/requirements.in
    #   django-filter
    #   djangorestframework
django-environ==0.7.0
    # via -r ../requirements/requirements.in
django-filter==21.1
    # via -r ../requirements/requirements.in
djangorestframework==3.12.4
    # via -r ../requirements/requirements.in
idna==3.2
    # via requests
psycopg2-binary==2.9.1
    # via -r ../requirements/requirements.in
pytz==2021.1
    # via django
requests==2.26.0
    # via -r ../requirements/requirements.in
sqlparse==0.4.2
    # via django
urllib3==1.26.7
    # via requests
